backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
boto3>=1.29.0
pgvector>=0.2.3
django-storages>=1.14.2
django-jazzmin>=2.6.0
whitenoise>=6.6.0
djangorestframework-simplejwt>=5.3.1 